+++ src/main/java/ca/sait/vezorla/model/AccountDiscount.java
@@ -8,7 +8,12 @@ import javax.persistence.*;
 import java.io.Serializable;
 
 /**
- * Account_Discount bridging table modeling
+ * AccountDiscount class.
+ *
+ * Account_Discount bridging table modeling.
+ *
+ * This class models a table within the database. As such,
+ * each variable represents a column.
  *
  * @author matthewjflee, jjrr1717
  */
@@ -20,11 +25,23 @@ import java.io.Serializable;
 @Table(name = "account_discount")
 public class AccountDiscount implements Serializable {
 
+    /**
+     * Account email.
+     *
+     * Annotated to be a Foreign Key for the database
+     * table 'account_discount'
+     */
     @Id
     @ManyToOne
     @JoinColumn(name="account_email")
     private Account email;
 
+    /**
+     * Discount code.
+     *
+     * Annotated to be a Foreign Key for the database
+     * table 'account_discount'
+     */
     @Id
     @ManyToOne
     @JoinColumn(name="discount_code")
